+++ javabackend/src/main/java/javabackend/example/javabackend/repositories/ordersRepository.java
@@ -3,7 +3,7 @@ package javabackend.example.javabackend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import javabackend.example.javabackend.models.*;
 import java.util.List;
-public interface ordersRepository extends CrudRepository<orders, Long> {
+public interface ordersRepository extends CrudRepository<orders, Integer> {
     List<orders> findAll();
 
 }
